+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+  * Reduce allocations in AV1 packetizer/depacketizer #889
   * Workaround DCEP panics due to reuse race condition #883
   * sprop-max-don-diff SDP parameter #885
   * Reduce number of RTP header extensions sent on every packet #880